\name{1.fasta}
\alias{1.fasta}
\alias{3.fasta}
\docType{data}
\title{Simulated SNP alignment data}
\description{Simulated SNP alignment files in FASTA format for use with the D-statistic functions.

\code{1.fasta} contains 4 sequences representing single individuals in the order P1, P2, P3, and Outgroup, for use with \code{\link{CalcD}} and \code{\link{WinCalcD}}.

\code{3.fasta} contains population-level data with multiple sequences per population (sequences from the same population share identical names), for use with \code{\link{CalcPopD}}.
}
\references{http://coleoguy.github.io/}
\author{Heath Blackmon}
\keyword{datasets}
