\name{data.taxonomy}
\alias{data.taxonomy}
\alias{tree.taxonomy}
\docType{data}
\title{taxonomic table}
\description{These files contain taxonomic information for the hymenoptera 
tree used in testing this package. Data was generated with the GetTaxonomy
 function.}
\references{http://www.uta.edu/karyodb/evobiR/}
\author{Heath Blackmon}
\keyword{datasets}
