\name{Even}
\alias{Even}
\title{
Tests whether a number is even}
\description{
Just a simple function that returns True if a number is even and False otherwise.}
\usage{
Even(x)
}
\arguments{
  \item{x}{a numerical vector.}
}
\details{
Returns a vector of logical values of the same length as the input vector.}
\references{
\url{http://www.uta.edu/karyodb/evobiR/}}
\author{
Heath Blackmon
}

\examples{
Even(c(1,2,3,4,5,6,2,5))
}
\keyword{ basic stats }
\keyword{ mode }
