\name{evobiR-package}
\alias{evobiR-package}
\alias{evobiR}
\docType{package}
\title{evobiR: Evolutionary Biology in R
}
\description{\pkg{evobiR} is a collection of tools for use in evolutionary biology. Functions include comparative methods for analyzing the relationship between continuous and discrete traits on phylogenies, implementations of Patterson's D-statistic for testing introgression, tools for manipulating sequence alignments and phylogenetic trees, and utility functions for sliding window analyses, residual selection, and population genetics calculations.}
\details{
\tabular{ll}{
Package: \tab evobiR\cr
Type: \tab Package\cr
Version: \tab 2.2\cr
Date: \tab 2025-02-20\cr
License: \tab GPL (>=2)\cr
}
More information on \pkg{evobiR} is available at \url{https://github.com/coleoguy/evobir}

}
\author{
Michelle M. Jonika, Maximos Chin, Nathan Anderson, Richard H. Adams, Jeffery P. Demuth, and Heath Blackmon

Maintainer: Heath Blackmon <coleoguy@gmail.com>
}
