\name{genome}
\alias{genome}
\docType{data}
\title{A small portion of the genome of Tribolium confusum}
\description{
This is just little piece of the Tribolium confusum genome
}
\format{
  a list of vectors with each element being a single string of the class "SeqFastadna".
}
\references{http://www.uta.edu/karyodb/evobiR/}
\author{Heath Blackmon}
\keyword{datasets}
