\name{getEJ}
\alias{getEJ}
\title{
creates an ej string}
\description{
Supplied with a phylo object it returns a character string describing a tree for ms.}
\usage{
getEJ(tree)
}
\arguments{
  \item{tree}{a phylo format tree}
}
\details{
Returns a character vector}
\references{
\url{http://coleoguy.github.io/}}
\author{
Heath Blackmon
}

\examples{
AICc(-32, 3, 100)
}
\keyword{ model comparison }
\keyword{ AICc }
