\name{hym.tree}
\alias{hym.tree}
\docType{data}
\title{Hymenoptera phylogenetic tree}
\description{A phylogenetic tree of class \code{phylo} containing 5 species of Hymenoptera (bees, wasps, and ants). Used in examples for the \code{\link{FuzzyMatch}} function to demonstrate fuzzy matching of taxon names between data and trees.}
\examples{
data(hym.tree)
ape::plot.phylo(hym.tree)
}
\keyword{datasets}
