\name{mite.data}
\alias{mite.data}
\alias{mite.tree}
\docType{data}
\title{cytogenetic and phylogenetic data for mites}
\description{This file contains cytogenetic mite data available at the tree of sex database as well as a phylogeny created from existing sequences from genbank.}
\references{http://www.uta.edu/karyodb/evobiR/}
\author{Heath Blackmon}
\keyword{datasets}
