% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pfsa.R
\name{print.PropSA}
\alias{print.PropSA}
\title{Print method for PropSA objects}
\usage{
\method{print}{PropSA}(x, ...)
}
\arguments{
\item{x}{A \code{PropSA} object returned by \code{\link{Prop_SA}}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for PropSA objects
}
