\name{taxa.table}
\alias{taxa.table}
\alias{species.tree}
\docType{data}
\title{data to illustrate high level tree function}
\description{These files contain a random tree for 18 insects as well as taxonomic information for each species in a table.}
\references{http://coleoguy.github.io/}
\author{Heath Blackmon}
\keyword{datasets}
