\name{trees.mite}
\alias{trees.mite}
\alias{trees}
\docType{data}
\title{10 phylogenetic trees for mites}
\description{A \code{multiPhylo} object containing 10 phylogenetic trees from a previously published study on mite sexual system evolution. These trees can be used with the associated \code{\link{mite.trait}} dataset for comparative analyses using \code{\link{AncCond}}. Loaded as the object \code{trees}.}
\examples{
data(trees.mite)
plot(trees[[1]])
}
\keyword{datasets}
