\name{trees.nex}
\alias{trees.nex}
\docType{data}
\title{100 simulated phylogenetic trees}
\description{A nexus-format file containing a collection of 100 simulated phylogenetic trees, each with 10 tips. Stored in the \code{inst/} directory. Used in examples for the \code{\link{SampleTrees}} function to demonstrate tree subsampling from posterior distributions.}
\keyword{datasets}
