\name{evobiR-package}
\alias{evobiR-package}
\alias{evobiR}
\docType{package}
\title{evobiR: Evolutionary Biology in R
}
\description{\pkg{evobiR} is a collection of tools for use in evolutionary biology.  Some of the functions manipulate data in a way not implemented by other functions while others calculate sequence statistics or perform simulations, either of data across trees or genetic and genomic simulations. }
\details{
\tabular{ll}{
Package: \tab evobiR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-08\cr
License: \tab GPL (>=2)\cr
}
More information on \pkg{evobiR} is available at \url{http://www.uta.edu/karyodb/evobiR/}

A full list of functions in the \pkg{evobiR} package is accesible by typing library(help = evobiR)
}
\author{
Heath Blackmon

Maintainer: Heath Blackmon <coleoguy@gmail.com>
}