\name{getPaths}
\alias{getPaths}
\title{
Get evolutionary pathways}
\description{
This function takes a tree of class phylo and returns all of the pathways from the root to the tips.}
\usage{
getPaths(tree)
}
\arguments{
  \item{tree}{a phylo object}
}
\references{
\url{http://coleoguy.github.io/}}
\author{
Heath Blackmon
}
\keyword{ tree manipulation }