\name{mite.trait}
\alias{mite.trait}
\docType{data}
\title{phenotype data for mites}
\description{A dataframe of sexual system and chromosome number data for mites.  The first column has species names, the second column has diploid numbers, and the third column contains 0 and 1 to indicate diplodiploidy or haplodiploidy sex determination.}
\keyword{datasets}