\name{scaleTreeRates}
\alias{scaleTreeRates}
\title{Phylogenetic analysis of heterogeneity in discrete character evolution}
\description{This function performs the phylogenetic methods for analysis of heterogeneity in rates of discrete character evolution described in Jonika et al. (2023). It estimates a scalar multiplier for each edge of the phylogeny that modifies the overall transition rate matrix, allowing rates to vary across the tree.}
\usage{
scaleTreeRates(tree, tip.states, model, fixedQ = NULL,
               max.ratio = 2, nbins = 10, max.transition = 1,
               var.start = FALSE, pi = "fitzjohn")
}
\arguments{
  \item{tree}{a tree of class \code{phylo}}
  \item{tip.states}{a named vector of tip states for some discrete character which is associated with the phylogeny. Order can differ from order of tips on phylogeny.}
  \item{model}{the model to be used for likelihood calculations. This can either be a \code{string} which can be passed to the \code{fitMk} function of \pkg{phytools} (e.g. \code{"ER"}, \code{"SYM"}, \code{"ARD"}) or a symmetrical transition \code{matrix} which has transitions between states categorized into distinct rate classes.}
  \item{fixedQ}{optional argument to be used when a Q-matrix with pre-estimated rates is available. Defaults to \code{NULL}, in which case rates are estimated using \code{fitMk}.}
  \item{max.ratio}{numeric value greater than one describing the maximum ratio of scalar bins to one. Defaults to 2, i.e. scalar bins range between 0.5 and 2.}
  \item{nbins}{integer giving the number of scalar bins above and below 1. Defaults to 10, i.e. 10 bins below 1 and 10 bins above 1 for a total of 21 bins inclusive of 1.}
  \item{max.transition}{integer giving the maximum number of bins which the scalar associated with an edge can differ from the scalar associated with its parent edge. Defaults to 1.}
  \item{var.start}{logical; whether or not to try all possible scalar values at the root of the tree. If \code{TRUE}, the analysis will be iterated across all possible root scalar values and the best tree (highest likelihood) returned. If \code{FALSE} (default), root scalar is set to one and only a single iteration is performed.}
  \item{pi}{string giving method to be used for estimating the root prior. Takes any option which can be passed to \pkg{phytools}' \code{fitMk}. Defaults to \code{"fitzjohn"}.}
}
\value{A phylogeny of class \code{phylo} and \code{phyloscaled}. Phylogeny has all elements normally included in an object of class \code{phylo}, with an additional element:
\item{scalar}{a \code{numeric} vector of scalars equal in length to the number of edges in phylogeny. Ordering of scalars is identical to the ordering of edges. Values greater than 1 indicate faster rates on that edge; values less than 1 indicate slower rates.}}
\seealso{
  \code{\link{plot.phyloscaled}}
}
\author{Maximos Chin and Heath Blackmon}
\references{http://coleoguy.github.io/}
\examples{
\donttest{
## Simple example with simulated data
library(ape)
library(phytools)
tree <- rtree(20)
tip.states <- sample(1:2, 20, replace = TRUE)
names(tip.states) <- tree$tip.label
result <- scaleTreeRates(tree, tip.states, model = "ER", nbins = 3, max.ratio = 2)
result$scalar
}
}
\keyword{comparative phylogenetics}
\keyword{discrete trait}
\keyword{rate heterogeneity}
